Verify declaration of input parameters
This will prevent users from forgetting to specify input parameters.
This is particularly useful for the `ssh-private-key-filename` parameter
which defines a default value that you could override with an empty
GitHub Actions secret.

The use of temporary environment variables in GitHub Actions is good
practice to protect the shell script from script injection attacks.
However, the `rsync-options` and `rsync-source-path` input parameters
rely on being inlined as arguments to the rsync command to make the
`--rsh` option and glob expansion work correctly.

Normally, GitHub Actions will provide environment variables for input
parameters automatically. However, this is currently not the case for
composite actions, thus requiring them to be defined manually.